+++ bs-config.json
@@ -1,0 +1,3 @@
+{
+	"startPath": "/sample/sample1-simple-usage/index.html"
+}
