+++ sample/sample1-simple-usage/index.html
@@ -3,20 +3,21 @@
     <meta>
     <meta charset="UTF-8"/>
     <title>ng2-accordion</title>
-    <link rel="stylesheet" href="../../node_modules/bootstrap/dist/css/bootstrap.css">
+    <link rel="stylesheet" href="https://npmcdn.com/bootstrap@3.3.7/dist/css/bootstrap.css">
 
-    <script src="../../node_modules/es6-shim/es6-shim.min.js"></script>
-    <script src="../../node_modules/zone.js/dist/zone.js"></script>
-    <script src="../../node_modules/reflect-metadata/Reflect.js"></script>
-    <script src="../../node_modules/systemjs/dist/system.src.js"></script>
+    <script src="https://npmcdn.com/es6-shim@0.35.1/es6-shim.js"></script>
+
+    <script src="https://npmcdn.com/zone.js@0.6.12/dist/zone.js"></script>
+    <script src="https://npmcdn.com/reflect-metadata@0.1.3/Reflect.js"></script>
+    <script src="https://npmcdn.com/systemjs@0.19.31/dist/system.js"></script>
 
     <script>
         System.config({
-            baseURL: "/ng2-accordion/", // replace with your baseURL here
+            baseURL: "localhost", // replace with your baseURL here
             defaultJSExtensions: true,
             map: {
-                "@angular": "node_modules/@angular",
-                "rxjs": "node_modules/rxjs"
+                "@angular": "/node_modules/@angular",
+                "rxjs": "/node_modules/rxjs"
             },
             packages: {
                 "app": { main: "main", defaultExtension: "ts" },
