--- src/index.ts
@@ -1,9 +1,0 @@
-export * from "./AccordionGroup";
-export * from "./Accordion";
-export * from "./AccordionToggle";
-
-import {AccordionGroup} from "./AccordionGroup";
-import {Accordion} from "./Accordion";
-import {AccordionToggle} from "./AccordionToggle";
-
-export const ACCORDION_DIRECTIVES: [any] = [AccordionGroup, Accordion, AccordionToggle];
